requirements.txt
streamlit>=1.28.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
langdetect>=1.0.9
tiktoken>=0.5.0